Resolved code smells.
